Try to use the SageMaker domain execution role to create the SageMaker project

+++ resources/lambdas/sagemaker_project/main.py
@@ -1,5 +1,7 @@
 import time
 import boto3
+import random
+import string
 from aws_lambda_powertools import Logger
 from crhelper import CfnResource
 from botocore.exceptions import ClientError
@@ -7,8 +9,8 @@ from typing import Dict
 
 helper = CfnResource()
 logger = Logger()
-sagemaker = boto3.client("sagemaker")
 catalog = boto3.client("servicecatalog")
+sts_connection = boto3.client('sts')
 
 
 @logger.inject_lambda_context(log_event=True)
@@ -18,8 +20,9 @@ def lambda_handler(event, context):
 @helper.create
 def create(event, _):
     project_properties = event.get("ResourceProperties")
-    project_name = project_properties["ProjectName"]
-    portfolio_id = project_properties["PortfolioId"]
+    project_name = project_properties.get("ProjectName")
+    portfolio_id = project_properties.get("PortfolioId")
+    domain_execution_role_arn = project_properties.get("DomainExecutionRoleArn")
     # Search for the SageMaker project product in the portfolio
     # We want to use the "MLOps template for model building, training, deployment and monitoring" product
     response = catalog.search_products_as_admin(
@@ -48,6 +51,7 @@ def create(event, _):
     latest_product_artifact_id = product_artifacts[0]["Id"]
     logger.info(f"Found latest product artifact: {product_artifacts[0]}")
     # Create the SageMaker Project for the SageMaker domain
+    sagemaker = get_sagemaker_client_with_domain_execution_role(domain_execution_role_arn)
     response = sagemaker.create_project(
         ProjectName=project_name,
         ProjectDescription="MLOps project for the Realtime Data Ingestion and Analytics solution",
@@ -74,23 +78,68 @@ def create(event, _):
 @helper.delete
 def delete(event, _):
     project_properties = event.get("ResourceProperties")
-    project_name = project_properties["ProjectName"]
-    # Delete the SageMaker project
-    response = sagemaker.delete_project(
-        ProjectName=project_name
-    )
-    # Wait for the SageMaker project to be deleted
-    deleted = False
-    while not deleted:
+    project_name = project_properties.get("ProjectName")
+    domain_execution_role_arn = project_properties.get("DomainExecutionRoleArn")
+    removal_policy = project_properties.get("RemovalPolicy", "destroy").lower()
+    if removal_policy == "destroy":
+        sagemaker = get_sagemaker_client_with_domain_execution_role(domain_execution_role_arn)
+        # Check if the SageMaker project exists.
+        # If it does not, just return as there is nothing to delete
         try:
-            sagemaker.describe_project(ProjectName=project_name)
+            response = sagemaker.describe_project(ProjectName=project_name)
         except ClientError as error:
-            if error.response["Error"]["Code"] == "ResourceNotFound":
-                logger.info(f"Deleted SageMaker project {project_name} successfully")
-                deleted = True
+            if error.response["Error"]["Code"] == "ValidationException":
+                logger.info(f"SageMaker project {project_name} does not exist")
                 return
-        time.sleep(5)
+        # Delete the SageMaker project
+        response = sagemaker.delete_project(
+            ProjectName=project_name
+        )
+        logger.info(f"Deleted SageMaker project: {response}")
+        # Wait for the SageMaker project to be deleted
+        deleted = False
+        while not deleted:
+            try:
+                sagemaker.describe_project(ProjectName=project_name)
+            except ClientError as error:
+                if error.response["Error"]["Code"] == "ValidationException":
+                    logger.info(f"Deleted SageMaker project {project_name} successfully")
+                    deleted = True
+                    return
+            time.sleep(5)
 
 @helper.update
 def do_nothing(_, __):
     logger.info("Nothing to do")
+
+def get_sagemaker_client_with_domain_execution_role(role_arn: str):
+    """This function assumes the domain execution role and use it to initialize the SageMaker client
+
+    Args:
+        role_arn (str): The domain execution role ARN
+
+    Returns:
+        SageMaker client
+    """
+    # Call the assume_role method of the STSConnection object and pass the role
+    # ARN and a role session name.
+    logger.info({f"Assuming the {role_arn} IAM Role to create the SageMaker Project"})
+    sts_response = sts_connection.assume_role(
+        RoleArn=role_arn,
+        RoleSessionName="SreateSagemakerProject"+"".join(random.choices(string.digits, k=10)),
+    )
+    logger.info({"sts AssumeRole response": sts_response})
+    # From the response that contains the assumed role, get the temporary 
+    # credentials that can be used to make subsequent API calls
+    credentials = sts_response["Credentials"]
+    access_key = credentials["AccessKeyId"]
+    secret_key = credentials["SecretAccessKey"]
+    session_token = credentials["SessionToken"]
+    # Initialize the SageMaker client using the assumed role credentials
+    sagemaker_client = boto3.client(
+        "sagemaker",
+        aws_access_key_id=access_key,
+        aws_secret_access_key=secret_key,
+        aws_session_token=session_token,
+    )
+    return sagemaker_client
